attaches seq# to wrap/unwrap
This attaches sequence number to wrap because in certain situations
of transform(wrap, wrap) it needs to create a node. In order to do
that it needs to know what the current sequence# from when the wraps
were created so it can createNode with the correct id.
Other transforms bring any lagged createNode ids up to date.