adds a weird scenario for T(wrap, wrap)
A situation where the wraps have intersecting children.
(Same parent of course.) Both have some outside of the other.
The idea is to have both effects take place. So imagine wraping nodes
with bold and another with italics.  In this case we need to clone
one of the two nodes and nest it under the other for the intersecting
parts.